Add section parsing
This is the first step in parsing a patch file. Section parsing splits
out a patch file into its different sections (each of which have their
own parsing logic) without attempting to parse each section.

The internal/text package included in this change is used throughout the
codebase as more tests are added.

Reviewed-on: D2385365