parse: Metavariables section
This sets up the skeleton for the program parser and adds support for
metavariable parsing. TODOs are left in place for parsing the actual
patch contents.

The metavariable section is parsed using the `go/scanner` package.

The internal/goast package included in this change makes it easier to
write tests that depend on token.Pos values where the base offset for a
file is uknown to the test case.

Reviewed-on: D2386373