parse/patch: Break apart unified diff
This commit begins the process of parsing the patch portion of a change.
The first step in the process is to split the unified diff apart into
the two versions of the code: before and after.

In future commits, we will parse these two versions. This commit only
goes as far as separating the unified diff.

Reviewed-on: D2387377