pgo: Parse into an AST
PGo is a superset of Go that we accept in the `-` or `+` section of a
gopatch patch.

This change defines a PGo AST, which is a superset of the Go AST, and
parses user-supplied code into it.

This is done by transforming the code into valid Go code (see D2387605
or HACKING.md to see how this is done) and relying on the standard
`go/parser` to actually parse the code. `go/parser` produces a standard
Go AST, which we traverse through, replacing nodes with PGo nodes where
necessary. This prevents us from having to write a custom Go parser.

This change closes the loop on PGo parsing. The same parsing logic is
used to parse either the `-` or the `+` section of a gopatch patch.

Reviewed-on: D2783405