Introduce engine
This sets us up to build the runtime engine for gopatch.

Once we have a parsed gopatch AST (defined in the internal/parse
package), we need to build a compiled representation of it. This is
where all the type-checking, etc. will take place.

This commit introduces a bare compiler that doesn't do anything except
copying the change name over.

Reviewed-on: D2870789