engine: Compile Meta section
Before we can patch any code, we need to understand the metavariables
defined in the `@@` section.

This change adds the Meta compiler which only understands the
`expression` and `identifier` types. At runtime, we only care about the
names and types of the metavariables so that's all we record.

At runtime, we only care about the names and types of the metavariables.
This will change in the future when we add support for constraints for
metavariable values.

The following planned but unimplemented features will affect this in the
future.

- Support for metavariables that reference user-defined types if the
  supplied Go code compiles successfully. This will require changing the
  MetavarType from an enum to a struct/interface.
- Support for metavariable constraints. These will allow us to specify
  generic matchers on metavariable values. This will require recording
  additional information for each metavariable besides the MetavarType.

Reviewed-on: D2870799