engine: Introduce Matcher
This introduces the Matcher abstraction. Matchers are compiled from the
minus section of a patch and are responsible for matching against Go
code.

For example, in the following diff,

    -foo()
    +bar()
    -baz()

The code,

    foo()
    baz()

Will be compiled into a Matcher.

We can then traverse a Go file and ask the complied Matcher if it
matches a Go file.

In terms of implementation, this change only includes a reflection-based
set of Matchers.

Future changes will include more matchers and tie them together with
file traversal.

Reviewed-on: D2907819