parse: Attach pgo files to AST
Now that we know how to parse PGo (our liberal superset of Go), we need
to attach the before and after versions to the patch AST before we can
proceed with actually interpreting the patches.

There's more to this than just using pgo.Parse on the two sides. When we
get a patch,

    -foo()
    +bar()

We generate two files:

    foo()       bar()

From the point of view of the PGo or Go parsers, these are two separate
files so they end up with different token.Pos base values in the parsed
AST. These token.Pos values are referencing positions in made up files
(.minus and .plus).

So after parsing the file, we have to inform the FileSet that different
offsets in these made up files map to actual positions in the original
patch file using [File.AddLineColumnInfo][1].

  [1]: https://golang.org/pkg/go/token/#File.AddLineColumnInfo

See the Position Tracking section of HACKING.md if the role of
token.Pos isn't clear.

This is the last piece we need in the parsing portion before we can
start interpreting patches.

Reviewed-on: D2870779