engine: Introduce Replacer
This introduces the Replacer abstraction. Replacers are compiled from
the plus section of a patch and are responsible for replacing matched
code.

For example, in the following diff,

    -foo()
    +bar()
    -baz()

The code,

    bar()

Will be compiled into a Replacer.

As we traverse the Go file with a Matcher, any matches it finds will be
replaced with the code from the Replacer.

This change only includes reflection-based replacers. Future changes
will include more replacers.

Reviewed-on: D2908045