engine: Matchers/Replacers information sharing
This introduces a Data object which acts as an immutable key-value
store similar with an API similar to context.Context and modifies the
Matcher and Replacer interfaces to share information using it.

    Match(reflect.Value, data.Data) (data.Data, ok bool)
    Replace(data.Data) (reflect.Value, error)

The idea is that matchers accumulate values into Data objects as they
traverse and match the AST. For example, if you're looking at an
ast.IfStmt, the matcher would do something along the lines of,

    d, ok = m.InitMatcher.Match(ifStmt.Init, d)
    if !ok { return d, ok }

    d, ok = m.CondMatcher.Match(ifStmt.Cond, d)
    if !ok { return d, ok }

    ...

Then when it's time to replace portions of the AST, replacers will use
information stored in the provided Data objects to build the new AST.

    &ast.IfStmt{
        Init: r.InitReplacer(d),
        Cond: r.CondReplacer(d),
        ...,
    }

Because our usage divides the workload into write-heavy and read-heavy
versions, unlike `context.Context`, Data includes the ability to index
all values for fast lookups.

So at a higher level, you'd do something similar to,

    d := change.Match(file)
    d = data.Index(d)
    file = change.Replace(file)

Reviewed-on: D2916203