engine: Metavariable support
This adds a Matcher and a Replacer, both of which together implement
support for metavariable in gopatch patches. This also demonstrates use
of data.Data objects (from the previous commit) to share information
between matchers and replacers.

Metavar Matchers are built for each occurrence of a metavariable in the
minus section of the patch and Replacers for each occurrence in the plus
section. Matchers store match data on the Data objects, which is later
retrieved by Replacers.

Reviewed-on: D2934079