engine: Add File Matcher and Replacer
This adds the File Matcher and Replacer types. These look similar to the
Matcher and Replacer interfaces but do not actually implement them.

These are the top-level Matchers and Replacers that operate on a parsed
Go file. Given a patch, the minus section becomes a FileMatcher, and the
plus section a FileReplacer.

This gives us roughly the interface,

    FileMatcher   func(*ast.File) data.Data
    FileReplacer  func(data.Data) *ast.File

Where FileMatcher and FileReplacer are stateless objects re-usable
across any number of input Go files. All match data for a file is stored
in the data.Data object, which the FileReplacer is able to produce an
ast.File from.

The diagram added to HACKING.md illustrates the flow of information.

Note that this change implements support for pgo.Expr (expression
transformations) only.

Reviewed-on: D3034431