engine/file: Support imports
This adds Support for adding and deleting imports from patches.

Handling of named imports and detection of import-usage is a bit iffy.
It'll work for the common cases but for drop in replacements like
net-context to context, it's currently treating the removed import as
"used". This last one should be fixable with relative ease in the
future.

Similarly, rewriting of named imports isn't handled. That is, if a patch
or a matched file used named imports, we need to learn how to rewrite
the identifers as well. This will require resolving package names and
changing identifier matching to rewrite as needed.

Reviewed-on: D3477613