Implement main and end-to-end tests
This implements the main.go for gopatch and sets up a basic end-to-end
test framework.

Note that the formatting in the test cases is all messed up. This is
expected behavior until we handle token.Pos values properly, which will
inform the Go formatter how we want the code formatted.

Reviewed-on: D3035871