engine: Add []Stmt Matcher and Replacer
This adds support for matching and replacing `[]Stmt` blocks anywhere in
a Go AST.

This required building a custom matcher and replacer pair because we
need to be able to match three kinds of AST nodes: BlockStmt,
CaseClause, and CommClause. Upon finding a match, we record the kind of
node we encountered and all its original fields in a data.Data object.
This is later used by the Replacer to reproduce the original value, with
only the statements changed.

Note that this does not support elision ('...').

Reviewed-on: D3051379