engine/file: Support GenDecl
GenDecl is a generic declaration in Go accounting for type, variable,
const, and import declarations. For the first three, it's used for both,
top-level declarations and for statement-level declarations.

This change adds some very basic support for GenDecl matching and
replacing in gopatch. One of the tests revealed a minor oversight in the
pgo Parser which is also fixed in this change.

Caveats:

- Paretheses (`var (foo string)`) aren't handled well because we're
  ignoring `token.Pos` values. The validity of `token.Pos` values
  informs the AST whether there are Paretheses or not.
- It's an all-or-nothing match. `var foo string` will only match
  `var foo string`, and not the `foo string` of `var foo, bar string` or
  of `var (foo string; bar string)`. We'll want to fix this in the
  future.

Reviewed-on: D3044073