e2e_test: Rename .in.go files to .go
This changes the e2e_test implementation to rename .in.go files to .go,
and not project the .out.go files out to disk at all.

Previously, we were writing both, foo.in.go and foo.out.go to a
temporary directory, transforming in.go, and checking its contents
against the in-memory version of out.go.

Besides just writing files we don't need to-disk, this had a negative
side-effect that tests didn't have a lot of flexibility in their file
names.

This changes the e2e test to rewrite .in.go files to .go, and omit the
.out.go files from being projected to disk because their contents are
already in-memory.

This ended up simplifying the implementation of the test a little as
well because it required separating temporary directory creation from
test loading.

Reviewed-on: D3202741