Match _test.go files
As part of a refactor, users wil want to transform both, application
code and tests for it.

Previously, gopatch was ignoring all test files because `go/packages`
did not yield them by default in the result of `packages.Load`.

Unfortunately, including tests is not as simple as setting the `Tests`
flag on `go/packages` due to a limitation in `go list`. `go list` does
not accept `-test` and `-find` at the same time.

	$ go list -test -find ./...
	go list -test cannot be used with -find

This is a problem because without `-find`, `go list` expects the code to
compile/the imports to be present on-disk, which explicitly goes against
gopatch's requirement of being able to work on code that doesn't
compile.

To work around this, we'll enumerate `*_test.go` files inside the source
directories.

Caveat: This will not work for directories that contain *just* test
files. Since the workaround for that is to add a minimal .go file to the
directory, fixing this is going to be low priority for now.

Reviewed-on: D3202763