engine: Match for and range with "for ..."
The Go AST differentiates between for statements and range statements.

For statements, defined by [`ast.ForStmt`][1] refer to statements in the
form,

	for init; cond; step {
		....
	}

  [1]: https://golang.org/pkg/go/ast/#ForStmt

Where `init` and `step` are optional. If just `cond` is specified, it's
equivalent to a `while` loop.

	for cond {
		...
	}

Range statements, defined by [`ast.RangeStmt`][2] refer to statements in the
form,

	for k, v := range x {
		...
	}

Where `k` and `v` are optional, `:=` may be `=`, and `x` is a map,
slice, or a channel.

  [2]: https://golang.org/pkg/go/ast/#RangeStmt

Most users won't care about this distinction. When you say `for ...` in
a patch, you likely want to match all loops, regardless of whether they
are For or Range statements.

This patch adds support for doing so by treating "for ..." as a special
matcher/replacer combo.

When "for ..." is encountered, provided it appeared on a context line
(no preceding `-` or `+`), it will match both, For and Range statements,
and recreate the original statement as-is, running the transformation
only on the body.

Reviewed-on: D3514783