Add workaround for pgo type mismatch
If a patch has a single line on one side that could be interpreted as a
single expression, and a statement on the other side, then the patch
fails to apply.

Add a hack to balance the level we believe the patch applies at by
making sure they're both statements if one of them is a statement.

Reviewed-on: D5065475