Add failing test case and t.Skip logic
During the migration of wayfare from the `pricing/to` library
to the `thriftrw/ptr` library, we discovered a bug in import
handling.

As a more effective and informative "todo" list, it was decided
to add a failing test case for the bug, and logic to skip such
test cases in the `e2e_test.go` integration tests.

Reviewed-on: D4855977