engine/change: Connect the dots
In order to better match dots from the minus section of a patch to their
counterparts in the plus section, track all dots collected during matching
and replacing and "connect" them when compiling the Change.

While here, refactor to use the token.Pos directly for the `forDotsKey` now
that we no longer need line and column numbers to match correctly.

Reviewed-on: D4948869
Co-authored-by: Syrie Bianco <syrie@uber.com>