engine/file: Assert type name unchanged
For a patch like,

	-foo
	+foo.Bar

We want to replace all instances of `foo` with `foo.Bar` **except** when `foo`
is just a plain identifier, like a field of a struct or a variable name.

For example, in the following, we don't want `foo` to be replaced with `foo.Bar`.

	var foo int

This adds a test to verify this behavior, which was fixed in a previous diff.

Reviewed-on: D5031893
Co-authored-by: Syrie Bianco <syrie@uber.com>