Fix unintentional deletion of unchanged imports
astutil.AddNamedImport returns whether it actually added the import or
not, which was previously being ignored. This caused the system to treat
unchanged imports as newly added, in which case it determined that it
was okay to delete the old copy. This resulted in those imports getting
deleted unintentionally.

Fix this issue by actually inspecting the result of AddNamedImport, and
not marking the import as added when it wasn't added.

The included test case fails without this change.

Note: Whitespace handling isn't perfect for field deletion so the
included test case doesn't behave quite right.

Reviewed-on: D4963615