test/e2e: Support testing examples
While trying to add examples to the README, I ran into cases where I wanted
to have automated tests for them. Rather than point users to testdata/ for
examples, or try to keep examples/ in sync with testdata/, I decided it
made sense to add support for testing examples directly.

To test examples, the same format is used as for other end-to-end tests,
except that a patch body can now be,

```
=> path/to/another/file.patch
```

Where the path is relative to the location of e2e_test.