import/metavar: Recognize empty imports
Users previously had to work around the limitation that a named import in a
patch would transform an unnamed import to a named import.

That is,

```diff
@@
var foo, X identifier
@@
-import foo "example.com/foo"
+import foo "example.com/foo/v2"

 foo.X
```

Would transform unnamed imports in matched files to named imports:

```diff
package thing

-import "example.com/foo"
+import foo "example.com/foo/v2"
```

This was undesirable.

Teach the import matcher and replacer to recognize that the matched import
is unnamed, and remember that for later, associating it with the
metavariable name, not the import path. Associating it with the import path
would make it impossible to rename import paths like the case above.

The implementaton is a bit messy, but we basically have to differentiate
between the package name of an import and the identifier we use to refer to
types defined in that import.

Resolves #2