import/unnamed: Avoid duplicating
When rewriting unnamed imports like so, gopatch would sometimes fail to
delete the old import because it didn't realize that that import was no
longer used.

```
@@
var X identifier
@@
-import "example.com/foo/internal/lib"
+import "example.com/foo/lib"

  lib.X
```

This is because when detecting whether the import has been replaced, it
recorded the package name for the old import as unknown.

This changes it to guess the package name for the old import based on its
basename.

Caveat: Like the other place where we guess package names for unnamed
imports, this won't work when the package name doesn't match basename. We
can write a best effort heuristic for this, but I think our recommendation
here more generally should be to recommend a patch with a named import
metavariable so that we can handle both, named and unnamed imports, and
have a good idea of the package name from the patch.