Set up goreleaser
Set up goreleaser to publish binary releases to GitHub automatically.

The way this works is, when we push a new tag, it will use the
previously added extract-changelog tool to extract the release notes for
that release. It will then use goreleaser to publish a release.

goreleaser stamps the version number onto the binary with the following
ldflag:

    -X main._version={{.Version}}

This requires `_version` to be a `var`.