main/e2e: Readability improvements (#72)
Turn main's `run(...)` with N args into a `mainCmd` struct,
that has fields for global/OS-level state like stdin/stdout, etc.
and a Run method to accept CLI arguments to run with.

Use this struct to build a helper inside e2e_test
that runs main with provided arguments
and returns results to make the assertions inside the test easier.