diff: Use relative paths if input was relative (#73)
For the `--diff` flag, if the input file path was relative,
also use relative file paths in generated diffs.

Besides providing more readable output,
this makes the output of `--diff` more useful
because these are now effectively relocatable .patch files.

This change also allows us to drop the hack in e2e test
where we drop the first two lines of the `--diff` output
so that we have something deterministic to compare against.