Add -P / --patches-file flag
This adds a new flag '-P' that accepts a path to a file.
This file is expected to contain paths to zero or more other files,
which are all interpreted as arguments to '-p'.

So the following,

    % cat > patches.txt < EOF
    foo.patch
    bar.patch
    baz.patch
    EOF

    % gopatch -P patches.txt ...

Is equivalent to,

    % gopatch -p foo.patch -p bar.patch -p baz.patch ...

Empty lines in the patches file are ignored.
Comments are not supported at this time.

This makes it easier to write scripts where the list of patch files
comes from another source.

Refs GO-1679