This commit adds a flag `--skip-import-processing` to disables import formatting
for imports that were not part of the patch changes. (#110)
When the flag `--skip-import-processing` is set to true, we remove the call to
`imports.Process` which was formatting the imports after the patch was applied.
The flag being false has no affect and follows original behavior.

This was tested in the testcase testdata/noop_import, where without the
flag we get default patching. When the flag is set to true, the e2e_test
confirms the diff doesn't have group ordering applied.

An additional file suffix `.groupimports` is added which contains the
diff between input file and the output after applying
`--skip-import-processing` flag.

Internal Ref: GO-1955