Fix path for the testdata file `diff_example/error.go` (#151)
This PR adds the missing example
`testdata/test_files/diff_example/error.go` that is mentioned in the
documentation. This file has been mistakenly committed as
`testdata/test_files/lint_example/test2.go` in PR #64.

["Apply the patch"](https://github.com/uber-go/gopatch#apply-the-patch)
section in README states:

> For example if we applied patch ~/s1028 to our testfile error.go
>  ```
> $ gopatch -d -p ~/s1028.patch ./testdata/test_files/diff_example/
> ```
> Output would be :
> 
> ```
> gopatch/testdata/test_files/diff_example/error.go:Replace redundant
fmt.Sprintf with fmt.Errorf
> --- gopatch/testdata/test_files/diff_example/error.go
> +++ gopatch/testdata/test_files/diff_example/error.go
> @@ -7,7 +7,7 @@
> 
> func foo() error {
>         err := errors.New("test")
> -       return errors.New(fmt.Sprintf("error: %v", err))
> +       return fmt.Errorf("error: %v", err)
> }
> 
>  func main() {
> 
> ```

But I checked the repo and this file is absent.

Additionally, I modified the patch `s1028` with the comment. So, the
output in README is consistent.